refactor AnonymousRecruitmentWarning section to use RSC
